Temporarily omit vignettes
Temporarily omit vignettes so that the existing action from the NG-CHM
repo can run. Background: since the last release from the NG-CHM
project, we have added a vignette to this NGCHMSupportFiles repo.
However, the action in the NG-CHM project that creates releases in
NGCHMSupportFiles does not include the packages build vignettes.
So the action failed when we released NG-CHM 2.24.4.

The plan is to disable the vignettes here, and rerun the action in the
NG-CHM project to create an NGCHMSupportFiles 2.24.4 release to include
the bug fixes in 2.24.4 (but w/o the vignette).

Then later, we'll update the NG-CHM action to be able to build
vignettes and also re-enable the vignettes here.